refactor: rename component and fix the importing
